requirements.txt
aiogram<3.0
aiohttp==3.8.1
aioredis==2.0.1
python-dotenv==0.20.0
